renames modules -> plugin
