+++ services/api-server/src/simcore_service_api_server/api/routes/studies.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends
 
 from ...core.settings import BasicSettings
 from ...models.schemas.studies import StudyID, StudyPort
-from ...modules.webserver import AuthSession
+from ...plugins.webserver import AuthSession
 from ..dependencies.webserver import get_webserver_session
 
 logger = logging.getLogger(__name__)
