+++ services/api-server/src/simcore_service_api_server/api/routes/users.py
@@ -3,7 +3,7 @@ import logging
 from fastapi import APIRouter, Depends, Security
 
 from ...models.schemas.profiles import Profile, ProfileUpdate
-from ...modules.webserver import AuthSession
+from ...plugins.webserver import AuthSession
 from ..dependencies.webserver import get_webserver_session
 
 logger = logging.getLogger(__name__)
