+++ services/api-server/tests/unit/test_models_schemas_files.py
@@ -13,7 +13,7 @@ from fastapi import UploadFile
 from models_library.api_schemas_storage import FileMetaDataGet as StorageFileMetaData
 from pydantic import ValidationError
 from simcore_service_api_server.models.schemas.files import File
-from simcore_service_api_server.modules.storage import to_file_api_model
+from simcore_service_api_server.plugins.storage import to_file_api_model
 
 FILE_CONTENT = "This is a test"
